授权和支付

+++ src/main/java/com/zx/model/OrderMaster.java
@@ -49,13 +49,13 @@ public class OrderMaster {
      * 订单状态, 默认为新下单
      */
     @TableField(value = "order_status")
-    private Byte orderStatus;
+    private Integer orderStatus;
 
     /**
      * 支付状态, 默认未支付
      */
     @TableField(value = "pay_status")
-    private Byte payStatus;
+    private Integer payStatus;
 
     /**
      * 创建时间
